requirements.txt
antlr4-python3-runtime==4.9.2
click~=8.0.1
mingus~=0.6.1
pygame~=2.0.1